+++ inst/fw_templ/p_blank/ui_header.R
@@ -52,7 +52,8 @@ left_menu <- NULL
 right_menu <- NULL
 
 # -- Register Header Elements in the ORDER SHOWN in the UI
-add_ui_header(left_menu          = left_menu,
+add_ui_header(title              = "Set using add_ui_header() in program/ui_header.R",
+              left_menu          = left_menu,
               right_menu         = right_menu,
               skin               = skin,
               status             = status,