+++ tests/testthat/sample_app_left_sidebar/program/ui_header.R
@@ -52,7 +52,8 @@ left_menu <- NULL
 right_menu <- NULL
 
 # -- Register Header Elements in the ORDER SHOWN in the UI
-add_ui_header(left_menu          = left_menu,
+add_ui_header(title              = "periscope2 Example Application",
+              left_menu          = left_menu,
               right_menu         = right_menu,
               skin               = skin,
               status             = status,