+++ tests/testthat/test_download_file.R
@@ -30,6 +30,10 @@ download_string_list <- function() {
     c("test1", "test2", "tests")
 }
 
+download_char_data <- function() {
+    "A123B"
+}
+
 # UI Testing
 test_that("downloadFileButton", {
     file_btn <- downloadFileButton(id            = "myid",
@@ -37,6 +41,13 @@ test_that("downloadFileButton", {
                                    hovertext     = "myhovertext")
     expect_true(grepl('title="myhovertext"', file_btn, fixed = TRUE))
     expect_true(grepl('id="myid-csv"', file_btn, fixed = TRUE))
+})
+
+test_that("downloadFileButton - no download type", {
+    file_btn <- downloadFileButton(id            = "myid2",
+                                   downloadtypes = NULL,
+                                   hovertext     = "myhovertext")
+    expect_equal(file_btn, "")
 })
 
 test_that("downloadFileButton multiple types", {
@@ -132,9 +143,13 @@ test_that("downloadFile - download char data", {
     testServer(downloadFile,
                args = list(logger       = periscope2:::fw_get_user_log(),
                            filenameroot = "my_char_download",
-                           datafxns     = list(txt = function() {"123"})),
+                           datafxns     = list(txt = download_char_data,
+                                               tsv = download_char_data,
+                                               csv = download_char_data)),
                expr = {
                    expect_snapshot_file(output$txt)
+                   expect_snapshot_file(output$tsv)
+                   expect_snapshot_file(output$csv)
                })
 })
 
@@ -145,5 +160,25 @@ test_that("downloadFile - download txt numeric data", {
                            datafxns     = list(txt = function() {123})),
                expr = {
                    expect_warning(output$txt, "txt could not be processed")
+               })
+})
+
+test_that("downloadFile - default values", {
+    testServer(downloadFile,
+               args = list(datafxns = list(txt = function() {"123"})),
+               expr = {
+                   expect_snapshot_file(output$txt)
+               })
+})
+
+test_that("downloadFile - invalid type", {
+    testServer(downloadFile,
+               args = list(datafxns = list(ttt = function() {"123"},
+                                           jeg = download_lattice_plot,
+                                           tff = download_plot)),
+               expr = {
+                   expect_error(output$ttt)
+                   expect_error(output$jeg)
+                   expect_error(output$tff)
                })
 })